#1929 Configuration logic by provider (#1984)
* #1929 Configuration logic by provider

- Add CloudProvider dependencies to the Server Startup
- Add CloudProvider dependencies to the Infrastructure Startup
- Azure first in the switch to unify

* #1929 Healthcheck config

- Move common healthcheck configuration

* #1929 Move AWS Client

- Move AWS client singletons into the infrastructure Layer
