+++ src/AzureIoTHub.Portal.Infrastructure/AzureIoTHub.Portal.Infrastructure.csproj
@@ -106,6 +106,8 @@
   </ItemGroup>
 
   <ItemGroup>
+    <PackageReference Include="AWSSDK.IoT" Version="3.7.105.22" />
+    <PackageReference Include="AWSSDK.IotData" Version="3.7.102.5" />
     <PackageReference Include="Azure.Data.Tables" Version="12.8.0" />
     <PackageReference Include="Azure.Messaging.EventHubs" Version="5.8.1" />
     <PackageReference Include="Azure.Messaging.EventHubs.Processor" Version="5.8.1" />
